restore now available

+++ Smash/Scripts/Player.gd
@@ -6,6 +6,7 @@ var pos = 0
 var scorePos = 0
 var played = 0
 var mod = 1
+var beginPos = 0
 
 
 var new_position = Vector2(0, 0)
@@ -23,14 +24,15 @@ var playerBoxColor
 var playerBox
 var playerTween
 
+
 func _ready():
 	add_to_group("players")
 
 
 func _process(_delta):
-	if not fighting and pos < 4:
-		tween_player()
-	elif moving:
+	if pos < 4:
+		fighting = true
+	if moving:
 		if position.distance_to(new_position) > 10:
 			velocity = position.direction_to(new_position) * 700
 			move_and_slide()
@@ -40,16 +42,18 @@ func _process(_delta):
 			moving = false
 	elif pos > 3 and fighting or dnf:
 		fighting = false
-		playerTween.kill()
+		if playerTween:
+			playerTween.kill()
 
 
 func tween_player():
-	fighting = true
-	playerTween = get_tree().create_tween()
-	playerTween.tween_property(playerSprite, "position", Vector2(10*mod, 0), 0.25)
-	playerTween.tween_property(playerSprite, "position", Vector2(-20*mod, 0), 0.5)
-	playerTween.tween_property(playerSprite, "position", Vector2(10*mod, 0), 0.25)
-	playerTween.set_loops()
+	if fighting and pos < 4:
+		fighting = true
+		playerTween = get_tree().create_tween()
+		playerTween.tween_property(playerSprite, "position", Vector2(10*mod, 0), 0.25)
+		playerTween.tween_property(playerSprite, "position", Vector2(-20*mod, 0), 0.5)
+		playerTween.tween_property(playerSprite, "position", Vector2(10*mod, 0), 0.25)
+		playerTween.set_loops()
 
 func move(newPos):
 	new_position = newPos
@@ -88,6 +92,12 @@ func save_stats():
 		"pos": pos,
 		"scorePos": scorePos,
 		"playerSprite": playerSprite.texture.resource_path,
-		"playerName": playerName
+		"playerName": playerName,
+		"dnf": dnf,
+		"beginPos": beginPos
 	}
 	return save_dict
+	
+func kill_tween():
+	if playerTween:
+		playerTween.kill()
